:recycle: refactor(#168/onboarding): onboarding 도메인 관련 클래스 ninedot-domain 모듈로 이동
