+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/onboarding/exception/ChoiceException.java
@@ -1,7 +1,7 @@
 package org.sopt36.ninedotserver.onboarding.exception;
 
-import org.sopt36.ninedotserver.global.exception.BusinessException;
-import org.sopt36.ninedotserver.global.exception.ErrorCode;
+import org.sopt36.ninedotserver.exception.BusinessException;
+import org.sopt36.ninedotserver.exception.ErrorCode;
 
 public class ChoiceException extends BusinessException {
