+++ ninedot-domain/src/main/java/org/sopt36/ninedotserver/onboarding/model/Domain.java
@@ -1,0 +1,6 @@
+package org.sopt36.ninedotserver.onboarding.model;
+
+public enum Domain {
+    JOB,
+    PERSONA
+}
