--- src/main/java/org/sopt36/ninedotserver/onboarding/domain/Domain.java
@@ -1,6 +1,0 @@
-package org.sopt36.ninedotserver.onboarding.domain;
-
-public enum Domain {
-    JOB,
-    PERSONA
-}
